+++ api-admin/src/main/resources/api/openapi.yml
@@ -638,8 +638,8 @@ paths:
   /api/portals/v1/{portalShortcode}/studies/{studyShortcode}/env/{envName}/enrollee/search/v2/facets:
     get:
       summary: Gets all of the possible fields which can be searched on within the provided portal.
-      tags: [ enrolleeSearchExpression ]
-      operationId: getSearchFacets
+      tags: [ enrolleeSearch ]
+      operationId: getExpressionSearchFacets
       parameters:
         - { name: portalShortcode, in: path, required: true, schema: { type: string } }
         - { name: studyShortcode, in: path, required: true, schema: { type: string } }
