+++ core/src/main/java/bio/terra/pearl/core/service/survey/SurveyService.java
@@ -48,6 +48,10 @@ public class SurveyService extends VersionedEntityService<Survey, SurveyDao> {
         return dao.findByStableIdWithMappings(stableId, version);
     }
 
+    public List<Survey> findByStudyEnvironmentIdWithContent(UUID studyId) {
+        return dao.findByStudyEnvironmentIdWithContent(studyId);
+    }
+
     @Transactional
     @Override
     public void delete(UUID surveyId, Set<CascadeProperty> cascades) {
